Addressed MR comments. Renamed files, update timezone, message output prefix.
Have audit event log use MPF's time formatting.

fixing pattern
